requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
pytube==15.0.0
PyVimeo==1.1.0
requests==2.31.0
tinydb==4.8.0
tuspy==1.0.3
urllib3==2.2.1
yarl==1.9.4
